Spacy/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "spacy-intent"
version = "0.3.0"
description = "Classificador de intenções em português usando spaCy com configuração externa"
requires-python = ">=3.10"
dependencies = [
    "spacy>=3.7.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
postgres = [
    "psycopg2-binary>=2.9",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["spacy_intent*"]

[tool.setuptools.package-data]
spacy_intent = ["config/*.yml"]
